Update test_locale_filter.py
Fixes the test that makes sure filtering by locale works.

This nixes a test that was marked as xfail. I'm not sure it's that
interesting a test because I'm not sure we really care whether the two
numbers are almost the same. Plus since it's tickling bug #940361 in
Input which I'm not planning to fix anytime soon, I think we should just
rewrite it when the bug is fixed and thus carry less technical debt.

Fixes #147